deploy: add `-automaxprocs` to reduce CPU and memory resources
With the new `-automaxprocs` commandline parameter, Ceph-CSI will adjust
the GOMAXPROCS environment variable for the Golang runtime. The values
are based on the CPU quota that is given to the process. This can reduce
the number of threads that the Golang runtime spawns, which affects the
require amount of memory as well.

Updates: #5228
Signed-off-by: Niels de Vos <ndevos@ibm.com>