ci: add `ok-to-test` label after Mergify rephrase
Mergify rephrased the status when a PR is the 1st in the queue. Instead
of the "Queue: Embarked in merge train", the status is now reported as
"Queue: Embarked in merge queue".

Because the status did not match the rule in the configuration anymore,
the `ok-to-test` label was not automatically added.

Signed-off-by: Niels de Vos <ndevos@ibm.com>