doc: add `v3` symlink to get current versions in pkg.go.dev
https://pkg.go.dev/github.com/ceph/ceph-csi does not list the current
v3.14.0 release. It seems to come from the fact that a v3 release of a
Go module expects a v3/go.mod file in the repository. Instead of moving
all files from the root of the repository to a new `v3` directory, make
a symlink so that https://github.com/ceph/ceph-csi/v3/go.mod resolves
correctly (eventhough the go.mod does not reference /v3).

Signed-off-by: Niels de Vos <ndevos@ibm.com>