rbd: add check before calling Destroy
It seems possible that the .Destroy() function is
called on a nil pointer. This would cause a
panic in the node-plugin.

Depending on how far GenVolFromVolID() comes, the
returned rbdVolume can
be connected. If an error occurs, the
rbdVolume should not be connected
at all, so call the .Destroy() function in
those cases too.

Refactor the code to avoid complexity

Signed-off-by: Madhu Rajanna <madhupr007@gmail.com>
Co-authored-by: Niels de Vos <ndevos@ibm.com>
(cherry picked from commit 918f2549ef51c082ac18ba170a281f46257ae218)