util: always advertise Controller PUBLISH_UNPUBLISH cap
This commit changes to always advertise controller
PUBLICH_UNPUBLISH capability to avoid stale volumeattachments.

Previously, the Controller PUBLISH_UNPUBLISH capability was
only advertised when the 'enable-fencing' flag was set. In
this setup, the external-attacher would add a finalizer to
the VolumeAttachment. If the user later disabled fencing by
setting enable-fencing to false, the capability would no longer
be advertised, causing the external-attacher to skip finalizer
removal. This left the VolumeAttachment stuck in the deletion phase.

Signed-off-by: Praveen M <m.praveen@ibm.com>