nvmeof: prevent concurrent adding/removing of hosts to the gateway
When a host is added to or removed from the gateway, a lock protects
concurrent adding/removing of the same host in ControllerPublishVolume
or ControllerUnpublishVolume.

In case a 2nd gRPC operation is blocked due to the locking, an error is
returned to the caller. The caller will have to retry the operation at a
later time (usually Kubernetes retries with a very short delay).

Updates: #5370
Signed-off-by: Niels de Vos <ndevos@ibm.com>