rbd: add configuration option to use integrity verification mode
Problem: In a distributed settings it is not enough to establish data confidentiality with encryption but we also have verify the authenticity (integrity) of data.
In for example legislation like to EU's Digital Operational Resilience Act (DORA) confidently and authenticity of data has to be established.

Solution: By enabling the usage of the --integrity flag in dmcrypt we can use AEAD encryption mode which verifies the authenticity of read encrypted data.

Limitations: The luks resize operation is not compatible with using AEAD ciphers. Therefore if this option is chosen no resize is possible.
Signed-off-by: David Mohren <david.mohren@clyso.com>