e2e: add clone and snapshot tests with new encryption options
I looked into all  tests set the `"encrypted": true` and added a
variant where the cipher and integrity mode is set. Except for
tests that did a resize.

In addition, I added a new warning log entry to indicate that
resizing is not possible when the integrity flag is set.

Signed-off-by: David Mohren <david.mohren@clyso.com>