nfs: implement ControllerModifyVolume to change NFS-server hostname
The new hostname of the NFS-server is stored in a journal entry for the
volume. The new NFS-server will only be used the next time the NFS
Volume is mounted on a worker node.

It is not possible to update the VolumeContext of a Volume, so the old
hostname of the NFS=server will be stored there forever.

Closes: #5420
Signed-off-by: Niels de Vos <ndevos@ibm.com>